feat: move Url to sn_api
BREAKING CHANGE: Url is now in sn_api and not in sn
