+++ sn_api/src/app/mod.rs
@@ -30,9 +30,9 @@ pub mod multimap;
 pub mod nrs;
 pub mod register;
 pub mod resolver;
+pub use crate::url::*;
 pub use consts::DEFAULT_XORURL_BASE;
 pub use helpers::parse_tokens_amount;
-pub use safe_network::url::*;
 pub use xor_name::{XorName, XOR_NAME_LEN};
 
 #[derive(Clone)]
